requirements.txt
Django==2.2.0
psycopg2-binary==2.7.6.1
djangorestframework==3.8.2
djoser==1.3.2
djangorestframework-jwt==1.11.0
django-otp==0.5.1
django-countries==5.3.2
django-cors-headers==2.4.0
pycoin==0.80
requests==2.21.0
base58==1.0.3
mnemonic==0.18
gunicorn==19.9.0
gevent==1.3.7
pika==0.12.0
